# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handlers.CategoryEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.CategoryResponse'
      success:
        example: true
        type: boolean
    type: object
  handlers.CategoryListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.CategoryResponse'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  handlers.CourseEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.CourseResponse'
      success:
        example: true
        type: boolean
    type: object
  handlers.CourseListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.CourseResponse'
        type: array
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  handlers.CourseResponse:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      enrollment_capacity:
        type: integer
      enrollment_open:
        type: boolean
      id:
        type: string
      language:
        type: string
      level:
        type: string
      organization_id:
        type: string
      price_cents:
        type: integer
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      teacher_id:
        type: string
      teacher_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
      waitlist_enabled:
        type: boolean
    type: object
...
  handlers.MessageData:
    properties:
      message:
        type: string
      reset_token:
        type: string
    type: object
  handlers.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.MessageData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.createCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
      slug:
        type: string
    required:
    - name
    type: object
  handlers.createCourseRequest:
    properties:
      category_id:
        type: string
      cover_url:
        type: string
      currency:
        example: ETB
        type: string
      description:
        type: string
      language:
        example: en
        type: string
      level:
        example: beginner
        type: string
      organization_id:
        type: string
      price_cents:
        type: integer
      slug:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - title
    type: object
...
  handlers.setTagsRequest:
    properties:
      tags:
        items:
          type: string
        type: array
    required:
    - tags
    type: object
...
  handlers.updateCourseRequest:
    properties:
      category_id:
        type: string
      cover_url:
        type: string
      currency:
        type: string
      description:
        type: string
      language:
        type: string
      level:
        type: string
      price_cents:
        type: integer
      summary:
        type: string
      title:
        type: string
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@asnakech.com
    name: API Support
    url: http://www.asnakech.com/support
  description: |-
    Backend API for the Asnakech online education platform.
    All JSON responses use a shared envelope: `{ success, data, error, meta }`.
    Every response includes an `X-Request-ID` header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://www.asnakech.com/terms
  title: Asnakech School API
  version: 0.1.0
paths:
...
  /api/v1/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CategoryListEnvelope'
      summary: List categories
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Category
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CategoryEnvelope'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - courses
  /api/v1/courses:
    get:
      description: Public catalog defaults to published. Teachers/admins may see drafts
        they own (or all if admin).
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: per_page
        type: integer
      - description: Search
        in: query
        name: q
        type: string
      - description: Category slug
        in: query
        name: category
        type: string
      - description: Tag slug
        in: query
        name: tag
        type: string
      - description: Organization ID
        in: query
        name: organization_id
        type: string
      - description: Teacher ID
        in: query
        name: teacher_id
        type: string
      - description: beginner|intermediate|advanced
        in: query
        name: level
        type: string
      - description: draft|published|archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseListEnvelope'
      summary: List courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Course
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      security:
      - BearerAuth: []
      summary: Create course
      tags:
      - courses
  /api/v1/courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete course
      tags:
      - courses
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      summary: Get course
      tags:
      - courses
    patch:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Patch
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      security:
      - BearerAuth: []
      summary: Update course
      tags:
      - courses
...
  /api/v1/courses/{id}/archive:
    post:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      security:
      - BearerAuth: []
      summary: Archive course
      tags:
      - courses
...
  /api/v1/courses/{id}/publish:
    post:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      security:
      - BearerAuth: []
      summary: Publish course
      tags:
      - courses
  /api/v1/courses/{id}/tags:
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Tags
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.setTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CourseEnvelope'
      security:
      - BearerAuth: []
      summary: Replace course tags
      tags:
      - courses
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"